Added authentication and authorization

+++ Controllers/FountainPenOrderController.cs
@@ -2,6 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -10,6 +13,7 @@ using PenShop.Models;
 
 namespace PenShop.Controllers
 {
+    [Authorize]
     public class FountainPenOrderController : Controller
     {
         private readonly PenShopContext _context;
@@ -23,7 +27,23 @@ namespace PenShop.Controllers
         public async Task<IActionResult> Index()
         {
             var penShopContext = _context.FountainPenOrder.Include(f => f.Customer).Include(f => f.Order).Include(f => f.Pen);
-            return View(await penShopContext.ToListAsync());
+            var user = GetUser();
+            if (user is Administrator)
+            {
+                return View(await penShopContext.ToListAsync());
+            }
+            else if (user is Customer)
+            {
+                return View(await penShopContext.Where(x => x.CustomerId == user.Id || x.Order != null && x.Order.CustomerId == user.Id).ToListAsync());
+            }
+            else if (user is null)
+            {
+                return Unauthorized();
+            }
+            else
+            {
+                return Problem("Unknown user");
+            }
         }
 
         // GET: FountainPenOrder/Details/5
@@ -39,9 +59,16 @@ namespace PenShop.Controllers
                 .Include(f => f.Order)
                 .Include(f => f.Pen)
                 .FirstOrDefaultAsync(m => m.Id == id);
+
             if (fountainPenOrder == null)
             {
                 return NotFound();
+            }
+
+            var user = GetUser();
+            if (user is null || fountainPenOrder.CustomerId != user.Id && fountainPenOrder.Order?.CustomerId != user.Id && user is not Administrator)
+            {
+                return Unauthorized();
             }
 
             return View(fountainPenOrder);
@@ -50,11 +77,16 @@ namespace PenShop.Controllers
         // GET: FountainPenOrder/Create/penId
         public IActionResult Create(int penId)
         {
-            ViewData["CustomerId"] = new SelectList(_context.Customer, nameof(Customer.Id), nameof(Customer.FullName));
             var pen = _context.FountainPen.Where(x => x.Id == penId).FirstOrDefault();
             if (pen is null)
                 return NotFound();
 
+            var user = GetUser();
+            var customers = user is Customer
+                ? new Customer[] {(Customer)user}
+                : _context.Customer.AsEnumerable();
+
+            ViewData["CustomerId"] = new SelectList(customers, nameof(Customer.Id), nameof(Customer.FullName));
             ViewData["PenId"] = pen.Id;
             ViewData["PenName"] = pen.Name;
 
@@ -68,18 +100,30 @@ namespace PenShop.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(int penId, [Bind("RemoveNib,CustomerId,Quantity")] FountainPenOrder fountainPenOrder)
         {
+            var user = GetUser();
+            if (user is null || fountainPenOrder.CustomerId != user.Id && user is not Administrator)
+            {
+                return Unauthorized();
+            }
+
             fountainPenOrder.PenId = penId;
+
             if (ModelState.IsValid)
             {
                 _context.Add(fountainPenOrder);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["CustomerId"] = new SelectList(_context.Customer, nameof(Customer.Id), nameof(Customer.FullName), fountainPenOrder.CustomerId);
+
             var pen = _context.FountainPen.Where(x => x.Id == penId).FirstOrDefault();
             if (pen is null)
                 return NotFound();
 
+            var customers = user is Customer
+                ? new Customer[] {(Customer)user}
+                : _context.Customer.AsEnumerable();
+
+            ViewData["CustomerId"] = new SelectList(customers, nameof(Customer.Id), nameof(Customer.FullName), fountainPenOrder.CustomerId);
             ViewData["PenId"] = pen.Id;
             ViewData["PenName"] = pen.Name;
 
@@ -99,8 +143,23 @@ namespace PenShop.Controllers
             {
                 return NotFound();
             }
-            ViewData["CustomerId"] = new SelectList(_context.Customer, nameof(Customer.Id), nameof(Customer.FullName), fountainPenOrder.CustomerId);
-            ViewData["OrderId"] = new SelectList(_context.Order, nameof(Order.Id), nameof(Order.Text), fountainPenOrder.OrderId);
+
+            var user = GetUser();
+            if (user is null || fountainPenOrder.CustomerId != user.Id && fountainPenOrder.Order?.CustomerId != user.Id && user is not Administrator)
+            {
+                return Unauthorized();
+            }
+
+            var customers = user is Customer
+                ? new Customer[] {(Customer)user}
+                : _context.Customer.AsEnumerable();
+
+            var orders = user is Customer
+                ? _context.Order.Where(x => x.CustomerId == user.Id)
+                : _context.Order.AsEnumerable();
+
+            ViewData["CustomerId"] = new SelectList(customers, nameof(Customer.Id), nameof(Customer.FullName), fountainPenOrder.CustomerId);
+            ViewData["OrderId"] = new SelectList(orders, nameof(Order.Id), nameof(Order.Text), fountainPenOrder.OrderId);
             ViewData["PenId"] = new SelectList(_context.FountainPen, nameof(FountainPen.Id), nameof(FountainPen.Name), fountainPenOrder.PenId);
             return View(fountainPenOrder);
         }
@@ -112,16 +171,47 @@ namespace PenShop.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("PenId,RemoveNib,Id,CustomerId,OrderId,Quantity")] FountainPenOrder fountainPenOrder)
         {
+            var user = GetUser();
+            if (user is null)
+                return Unauthorized();
+
+            var customers = user is Customer
+                ? new Customer[] {(Customer)user}
+                : _context.Customer.AsEnumerable();
+
+            var orders = user is Customer
+                ? _context.Order.Where(x => x.CustomerId == user.Id)
+                : _context.Order.AsEnumerable();
+
+            if (!customers.Any(x => x.Id == fountainPenOrder.CustomerId) && !orders.Any(x => x.Id == fountainPenOrder.OrderId))
+            {
+                return Unauthorized();
+            }
+
             if (id != fountainPenOrder.Id)
             {
                 return NotFound();
             }
 
+            var oldFountainPenOrder = _context.FountainPenOrder
+                .Include(f => f.Customer)
+                .Include(f => f.Order)
+                .Include(f => f.Pen)
+                .FirstOrDefault(x => x.Id == id);
+            if(oldFountainPenOrder is null)
+                return NotFound();
+
+            oldFountainPenOrder.PenId = fountainPenOrder.PenId;
+            oldFountainPenOrder.RemoveNib = fountainPenOrder.RemoveNib;
+            oldFountainPenOrder.CustomerId = fountainPenOrder.CustomerId;
+            oldFountainPenOrder.OrderId = fountainPenOrder.OrderId;
+            oldFountainPenOrder.Quantity = fountainPenOrder.Quantity;
+
             if (ModelState.IsValid)
             {
                 try
                 {
-                    _context.Update(fountainPenOrder);
+                    //_context.Update(fountainPenOrder);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
@@ -137,8 +227,8 @@ namespace PenShop.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["CustomerId"] = new SelectList(_context.Customer, nameof(Customer.Id), nameof(Customer.FullName), fountainPenOrder.CustomerId);
-            ViewData["OrderId"] = new SelectList(_context.Order, nameof(Order.Id), nameof(Order.Text), fountainPenOrder.OrderId);
+            ViewData["CustomerId"] = new SelectList(customers, nameof(Customer.Id), nameof(Customer.FullName), fountainPenOrder.CustomerId);
+            ViewData["OrderId"] = new SelectList(orders, nameof(Order.Id), nameof(Order.Text), fountainPenOrder.OrderId);
             ViewData["PenId"] = new SelectList(_context.FountainPen, nameof(FountainPen.Id), nameof(FountainPen.Name), fountainPenOrder.PenId);
             return View(fountainPenOrder);
         }
@@ -161,6 +251,12 @@ namespace PenShop.Controllers
                 return NotFound();
             }
 
+            var user = GetUser();
+            if (user is null || fountainPenOrder.CustomerId != user.Id && fountainPenOrder.Order?.CustomerId != user.Id && user is not Administrator)
+            {
+                return Unauthorized();
+            }
+
             return View(fountainPenOrder);
         }
 
@@ -173,9 +269,16 @@ namespace PenShop.Controllers
             {
                 return Problem("Entity set 'PenShopContext.FountainPenOrder'  is null.");
             }
+
             var fountainPenOrder = await _context.FountainPenOrder.FindAsync(id);
             if (fountainPenOrder != null)
             {
+                var user = GetUser();
+                if (user is null || fountainPenOrder.CustomerId != user.Id && fountainPenOrder.Order?.CustomerId != user.Id && user is not Administrator)
+                {
+                    return Unauthorized();
+                }
+
                 _context.FountainPenOrder.Remove(fountainPenOrder);
             }
             
@@ -185,7 +288,15 @@ namespace PenShop.Controllers
 
         private bool FountainPenOrderExists(int id)
         {
-          return (_context.FountainPenOrder?.Any(e => e.Id == id)).GetValueOrDefault();
+            return (_context.FountainPenOrder?.Any(e => e.Id == id)).GetValueOrDefault();
+        }
+
+        private IdentityUser? GetUser(){
+            var userId = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if(userId is null)
+                return null;
+
+            return _context.Users.AsNoTracking().FirstOrDefault(x => x.Id == userId);
         }
     }
 }
