+++ Controllers/GeneralProductOrderController.cs
@@ -2,6 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -10,6 +13,7 @@ using PenShop.Models;
 
 namespace PenShop.Controllers
 {
+    [Authorize]
     public class GeneralProductOrderController : Controller
     {
         private readonly PenShopContext _context;
@@ -23,7 +27,23 @@ namespace PenShop.Controllers
         public async Task<IActionResult> Index()
         {
             var penShopContext = _context.GeneralProductOrder.Include(g => g.Customer).Include(g => g.Order).Include(g => g.GeneralProduct);
-            return View(await penShopContext.ToListAsync());
+            var user = GetUser();
+            if (user is Administrator)
+            {
+                return View(await penShopContext.ToListAsync());
+            }
+            else if (user is Customer)
+            {
+                return View(await penShopContext.Where(x => x.CustomerId == user.Id || x.Order != null && x.Order.CustomerId == user.Id).ToListAsync());
+            }
+            else if (user is null)
+            {
+                return Unauthorized();
+            }
+            else
+            {
+                return Problem("Unknown user");
+            }
         }
 
         // GET: GeneralProductOrder/Details/5
@@ -39,9 +59,16 @@ namespace PenShop.Controllers
                 .Include(g => g.Order)
                 .Include(g => g.GeneralProduct)
                 .FirstOrDefaultAsync(m => m.Id == id);
+
             if (generalProductOrder == null)
             {
                 return NotFound();
+            }
+
+            var user = GetUser();
+            if (user is null || generalProductOrder.CustomerId != user.Id && generalProductOrder.Order?.CustomerId != user.Id && user is not Administrator)
+            {
+                return Unauthorized();
             }
 
             return View(generalProductOrder);
@@ -50,11 +77,16 @@ namespace PenShop.Controllers
         // GET: GeneralProductOrder/Create/productId
         public IActionResult Create(int productId)
         {
-            ViewData["CustomerId"] = new SelectList(_context.Customer, nameof(Customer.Id), nameof(Customer.FullName));
             var product = _context.Product.Where(x => x.Id == productId).FirstOrDefault();
             if (product is null || product is FountainPen)
                 return NotFound();
 
+            var user = GetUser();
+            var customers = user is Customer
+                ? new Customer[] {(Customer)user}
+                : _context.Customer.AsEnumerable();
+
+            ViewData["CustomerId"] = new SelectList(customers, nameof(Customer.Id), nameof(Customer.FullName));
             ViewData["ProductId"] = product.Id;
             ViewData["ProductName"] = product.Name;
 
@@ -68,6 +100,12 @@ namespace PenShop.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(int productId, [Bind("Id,CustomerId,Quantity")] GeneralProductOrder generalProductOrder)
         {
+            var user = GetUser();
+            if (user is null || generalProductOrder.CustomerId != user.Id && user is not Administrator)
+            {
+                return Unauthorized();
+            }
+
             generalProductOrder.GeneralProductId = productId;
             if (ModelState.IsValid)
             {
@@ -79,6 +117,11 @@ namespace PenShop.Controllers
             if (product is null || product is FountainPen)
                 return NotFound();
 
+            var customers = user is Customer
+                ? new Customer[] {(Customer)user}
+                : _context.Customer.AsEnumerable();
+
+            ViewData["CustomerId"] = new SelectList(customers, nameof(Customer.Id), nameof(Customer.FullName), generalProductOrder.CustomerId);
             ViewData["ProductId"] = product.Id;
             ViewData["ProductName"] = product.Name;
 
@@ -98,8 +141,23 @@ namespace PenShop.Controllers
             {
                 return NotFound();
             }
-            ViewData["CustomerId"] = new SelectList(_context.Customer, nameof(Customer.Id), nameof(Customer.FullName), generalProductOrder.CustomerId);
-            ViewData["OrderId"] = new SelectList(_context.Order, nameof(Order.Id), nameof(Order.Text), generalProductOrder.OrderId);
+
+            var user = GetUser();
+            if (user is null || user.Id != generalProductOrder.CustomerId && generalProductOrder.Order?.CustomerId != user.Id && user is not Administrator)
+            {
+                return Unauthorized();
+            }
+
+            var customers = user is Customer
+                ? new Customer[] {(Customer)user}
+                : _context.Customer.AsEnumerable();
+
+            var orders = user is Customer
+                ? _context.Order.Where(x => x.CustomerId == user.Id)
+                : _context.Order.AsEnumerable();
+
+            ViewData["CustomerId"] = new SelectList(customers, nameof(Customer.Id), nameof(Customer.FullName), generalProductOrder.CustomerId);
+            ViewData["OrderId"] = new SelectList(orders, nameof(Order.Id), nameof(Order.Text), generalProductOrder.OrderId);
             ViewData["GeneralProductId"] = new SelectList(_context.Product.AsEnumerable().Select(x => x is FountainPen ? null : x).Where(x => x != null), nameof(Product.Id), nameof(Product.Name), generalProductOrder.GeneralProductId);
             return View(generalProductOrder);
         }
@@ -111,16 +169,47 @@ namespace PenShop.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("GeneralProductId,Id,CustomerId,OrderId,Quantity")] GeneralProductOrder generalProductOrder)
         {
+            var user = GetUser();
+            if (user is null)
+                return Unauthorized();
+
+            var customers = user is Customer
+                ? new Customer[] {(Customer)user}
+                : _context.Customer.AsEnumerable();
+
+            var orders = user is Customer
+                ? _context.Order.Where(x => x.CustomerId == user.Id)
+                : _context.Order.AsEnumerable();
+
+            if (!customers.Any(x => x.Id == generalProductOrder.CustomerId) && !orders.Any(x => x.Id == generalProductOrder.OrderId))
+            {
+                return Unauthorized();
+            }
+
             if (id != generalProductOrder.Id)
             {
                 return NotFound();
             }
 
+            var oldGeneralProductOrder = _context.GeneralProductOrder
+                .Include(g => g.Customer)
+                .Include(g => g.Order)
+                .Include(g => g.GeneralProduct)
+                .FirstOrDefault(x => x.Id == id);
+
+            if(oldGeneralProductOrder is null)
+                return NotFound();
+
+            oldGeneralProductOrder.GeneralProductId = generalProductOrder.GeneralProductId;
+            oldGeneralProductOrder.CustomerId = generalProductOrder.CustomerId;
+            oldGeneralProductOrder.OrderId = generalProductOrder.OrderId;
+            oldGeneralProductOrder.Quantity = generalProductOrder.Quantity;
+
             if (ModelState.IsValid)
             {
                 try
                 {
-                    _context.Update(generalProductOrder);
+                    //_context.Update(generalProductOrder);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
@@ -136,8 +225,8 @@ namespace PenShop.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["CustomerId"] = new SelectList(_context.Customer, nameof(Customer.Id), nameof(Customer.FullName), generalProductOrder.CustomerId);
-            ViewData["OrderId"] = new SelectList(_context.Order, nameof(Order.Id), nameof(Order.Text), generalProductOrder.OrderId);
+            ViewData["CustomerId"] = new SelectList(customers, nameof(Customer.Id), nameof(Customer.FullName), generalProductOrder.CustomerId);
+            ViewData["OrderId"] = new SelectList(orders, nameof(Order.Id), nameof(Order.Text), generalProductOrder.OrderId);
             ViewData["GeneralProductId"] = new SelectList(_context.Product.AsEnumerable().Select(x => x is FountainPen ? null : x).Where(x => x != null), nameof(Product.Id), nameof(Product.Name), generalProductOrder.GeneralProductId);
             return View(generalProductOrder);
         }
@@ -155,9 +244,16 @@ namespace PenShop.Controllers
                 .Include(g => g.Order)
                 .Include(g => g.GeneralProduct)
                 .FirstOrDefaultAsync(m => m.Id == id);
+
             if (generalProductOrder == null)
             {
                 return NotFound();
+            }
+
+            var user = GetUser();
+            if (user is null || generalProductOrder.CustomerId != user.Id && generalProductOrder.Order?.CustomerId != user.Id && user is not Administrator)
+            {
+                return Unauthorized();
             }
 
             return View(generalProductOrder);
@@ -172,9 +268,16 @@ namespace PenShop.Controllers
             {
                 return Problem("Entity set 'PenShopContext.GeneralProductOrder'  is null.");
             }
+
             var generalProductOrder = await _context.GeneralProductOrder.FindAsync(id);
             if (generalProductOrder != null)
             {
+                var user = GetUser();
+                if (user is null || generalProductOrder.CustomerId != user.Id && generalProductOrder.Order?.CustomerId != user.Id && user is not Administrator)
+                {
+                    return Unauthorized();
+                }
+
                 _context.GeneralProductOrder.Remove(generalProductOrder);
             }
             
@@ -185,6 +288,14 @@ namespace PenShop.Controllers
         private bool GeneralProductOrderExists(int id)
         {
           return (_context.GeneralProductOrder?.Any(e => e.Id == id)).GetValueOrDefault();
+        }
+
+        private IdentityUser? GetUser(){
+            var userId = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if(userId is null)
+                return null;
+
+            return _context.Users.AsNoTracking().FirstOrDefault(x => x.Id == userId);
         }
     }
 }
